requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
evaluate>=0.4.0
sacrebleu>=2.3.0
rouge-score>=0.1.2
sentencepiece>=0.1.99
protobuf>=3.20.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0